Add initial pytorch policy
